+++ paymentsheet/src/main/java/com/stripe/android/customersheet/StripeCustomerAdapter.kt
@@ -25,10 +25,11 @@ import kotlin.coroutines.CoroutineContext
  */
 @OptIn(ExperimentalCustomerSheetApi::class)
 @JvmSuppressWildcards
-internal class StripeCustomerAdapter @Inject constructor(
+internal class StripeCustomerAdapter @Inject internal constructor(
     private val context: Context,
     private val customerEphemeralKeyProvider: CustomerEphemeralKeyProvider,
     private val setupIntentClientSecretProvider: SetupIntentClientSecretProvider?,
+    override val paymentMethodTypes: List<String>?,
     private val timeProvider: () -> Long,
     private val customerRepository: CustomerRepository,
     private val prefsRepositoryFactory: (CustomerEphemeralKey) -> PrefsRepository,
@@ -42,16 +43,32 @@ internal class StripeCustomerAdapter @Inject constructor(
         get() = setupIntentClientSecretProvider != null
 
     override suspend fun retrievePaymentMethods(): CustomerAdapter.Result<List<PaymentMethod>> {
+        paymentMethodTypes?.filter { PaymentMethod.Type.fromCode(it) == null }?.takeIf { it.isNotEmpty() }?.run {
+            val cause = IllegalStateException("Invalid payment method types provided (${joinToString()}).")
+            return CustomerAdapter.Result.failure(cause, null)
+        }
+
+        val supportedPaymentMethodCodes = supportedPaymentMethodTypes.map { it.code }.toSet()
+        val unsupportedPaymentMethods = paymentMethodTypes?.minus(supportedPaymentMethodCodes) ?: emptyList()
+        if (unsupportedPaymentMethods.isNotEmpty()) {
+            val unsupportedCodes = unsupportedPaymentMethods.joinToString()
+            val cause = IllegalStateException("Unsupported payment method types provided ($unsupportedCodes).")
+            return CustomerAdapter.Result.failure(cause, null)
+        }
+
+        val requestedTypes = if (paymentMethodTypes.isNullOrEmpty()) {
+            supportedPaymentMethodTypes
+        } else {
+            paymentMethodTypes.mapNotNull { PaymentMethod.Type.fromCode(it) }
+        }
+
         return getCustomerEphemeralKey().map { customerEphemeralKey ->
             customerRepository.getPaymentMethods(
                 customerConfig = PaymentSheet.CustomerConfiguration(
                     id = customerEphemeralKey.customerId,
                     ephemeralKeySecret = customerEphemeralKey.ephemeralKey,
                 ),
-                types = listOf(
-                    PaymentMethod.Type.Card,
-                    PaymentMethod.Type.USBankAccount,
-                ),
+                types = requestedTypes,
                 silentlyFail = false,
             ).getOrElse {
                 return CustomerAdapter.Result.failure(
@@ -190,6 +207,11 @@ internal class StripeCustomerAdapter @Inject constructor(
     internal companion object {
         // 30 minutes, server-side timeout is 60
         internal const val CACHED_CUSTOMER_MAX_AGE_MILLIS = 60 * 30 * 1000L
+
+        private val supportedPaymentMethodTypes: List<PaymentMethod.Type> = listOf(
+            PaymentMethod.Type.Card,
+            PaymentMethod.Type.USBankAccount,
+        )
     }
 }